+++ dev/org-ml-examples.el
@@ -1349,146 +1349,146 @@
            (org-ml-to-trimmed-string))
       => "#+call: cthulhu[:cache no](x=4) :exports results"
 
-      ;; :begin-hidden
-      ;; (:buffer "CLOCK: [2019-01-01 Tue]")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property
-      ;;       :value (org-ml-build-timestamp! '(2019 1 1) :end '(2019 1 2)))
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "CLOCK: [2019-01-01 Tue]--[2019-01-02 Wed] => 24:00"
+      :begin-hidden
+      (:buffer "CLOCK: [2019-01-01 Tue]")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property
+            :value (org-ml-build-timestamp! '(2019 1 1) :end '(2019 1 2)))
+           (org-ml-to-trimmed-string))
+      => "CLOCK: [2019-01-01 Tue]--[2019-01-02 Wed] => 24:00"
 
-      ;; (:buffer "~learn to~")
-      ;; (->> (org-ml-parse-this-object)
-      ;;      (org-ml-set-property :value "why?")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "~why?~"
+      (:buffer "~learn to~")
+      (->> (org-ml-parse-this-object)
+           (org-ml-set-property :value "why?")
+           (org-ml-to-trimmed-string))
+      => "~why?~"
 
-      ;; (:buffer "# not here")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :value "still not here")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "# still not here"
+      (:buffer "# not here")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :value "still not here")
+           (org-ml-to-trimmed-string))
+      => "# still not here"
 
-      ;; (:buffer "#+begin_comment"
-      ;;          "not here"
-      ;;          "#+end_comment")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :value "still not here")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => (:result "#+begin_comment"
-      ;;             "still not here"
-      ;;             "#+end_comment")
+      (:buffer "#+begin_comment"
+               "not here"
+               "#+end_comment")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :value "still not here")
+           (org-ml-to-trimmed-string))
+      => (:result "#+begin_comment"
+                  "still not here"
+                  "#+end_comment")
 
-      ;; (:buffer "%%(print :valueble)")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :value '(print :invaluble))
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "%%(print :invaluble)"
+      (:buffer "%%(print :valueble)")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :value '(print :invaluble))
+           (org-ml-to-trimmed-string))
+      => "%%(print :invaluble)"
 
-      ;; (:buffer ":LOGBOOK:"
-      ;;          ":END:")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :drawer-name "BOOKOFSOULS")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => (:result ":BOOKOFSOULS:"
-      ;;             ":END:")
+      (:buffer ":LOGBOOK:"
+               ":END:")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :drawer-name "BOOKOFSOULS")
+           (org-ml-to-trimmed-string))
+      => (:result ":BOOKOFSOULS:"
+                  ":END:")
 
-      ;; (:buffer "#+begin: blockhead"
-      ;;          "#+end:")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :block-name "blockfoot")
-      ;;      (org-ml-set-property :arguments '(:cache no))
-      ;;      (org-ml-to-trimmed-string))
-      ;; => (:result "#+begin: blockfoot :cache no"
-      ;;             "#+end:")
+      (:buffer "#+begin: blockhead"
+               "#+end:")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :block-name "blockfoot")
+           (org-ml-set-property :arguments '(:cache no))
+           (org-ml-to-trimmed-string))
+      => (:result "#+begin: blockfoot :cache no"
+                  "#+end:")
 
-      ;; (:buffer "\\pi")
-      ;; (->> (org-ml-parse-this-object)
-      ;;      (org-ml-set-property :name "gamma")
-      ;;      (org-ml-set-property :use-brackets-p t)
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "\\gamma{}"
+      (:buffer "\\pi")
+      (->> (org-ml-parse-this-object)
+           (org-ml-set-property :name "gamma")
+           (org-ml-set-property :use-brackets-p t)
+           (org-ml-to-trimmed-string))
+      => "\\gamma{}"
 
-      ;; ;; TODO test preserve indentation...
-      ;; (:buffer "#+begin_example"
-      ;;          "#+end_example")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :switches '("-n"))
-      ;;      (org-ml-set-property :value "example.com")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => (:buffer "#+begin_example -n"
-      ;;             "  example.com"
-      ;;             "#+end_example")
+      ;; TODO test preserve indentation...
+      (:buffer "#+begin_example"
+               "#+end_example")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :switches '("-n"))
+           (org-ml-set-property :value "example.com")
+           (org-ml-to-trimmed-string))
+      => (:buffer "#+begin_example -n"
+                  "  example.com"
+                  "#+end_example")
 
-      ;; (:buffer "#+begin_export latex"
-      ;;          "#+end_export")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :type "domestic")
-      ;;      (org-ml-set-property :value "bullets, bombs, and bigotry")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => (:buffer "#+begin_export domestic"
-      ;;             "bullets, bombs, and bigotry"
-      ;;             "#+end_export")
+      (:buffer "#+begin_export latex"
+               "#+end_export")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :type "domestic")
+           (org-ml-set-property :value "bullets, bombs, and bigotry")
+           (org-ml-to-trimmed-string))
+      => (:buffer "#+begin_export domestic"
+                  "bullets, bombs, and bigotry"
+                  "#+end_export")
 
-      ;; (:buffer "@@back-end:value@@")
-      ;; (->> (org-ml-parse-this-object)
-      ;;      (org-ml-set-property :back-end "latex")
-      ;;      (org-ml-set-property :value "new-value")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "@@latex:new-value@@"
+      (:buffer "@@back-end:value@@")
+      (->> (org-ml-parse-this-object)
+           (org-ml-set-property :back-end "latex")
+           (org-ml-set-property :value "new-value")
+           (org-ml-to-trimmed-string))
+      => "@@latex:new-value@@"
 
-      ;; (:buffer ": fixed")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :value "unfixed")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => ": unfixed"
+      (:buffer ": fixed")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :value "unfixed")
+           (org-ml-to-trimmed-string))
+      => ": unfixed"
 
-      ;; (:buffer "[fn:whitelabel] society")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :label "blacklabel")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "[fn:blacklabel] society"
+      (:buffer "[fn:whitelabel] society")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :label "blacklabel")
+           (org-ml-to-trimmed-string))
+      => "[fn:blacklabel] society"
 
-      ;; (:buffer "* dummy"
-      ;;          "stuff")
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :archivedp t)
-      ;;      (org-ml-set-property :commentedp t)
-      ;;      (org-ml-set-property :level 2)
-      ;;      (org-ml-set-property :pre-blank 1)
-      ;;      (org-ml-set-property :priority ?A)
-      ;;      (org-ml-set-property :tags '("tmsu"))
-      ;;      (org-ml-set-property :title '("smartie"))
-      ;;      (org-ml-set-property :todo-keyword "TODO")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => (:result "** TODO COMMENT [#A] smartie :tmsu:ARCHIVE:"
-      ;;             ""
-      ;;             "stuff")
-      ;; :begin-hidden
-      ;; (->> (org-ml-parse-this-element)
-      ;;      (org-ml-set-property :footnote-section-p t)
-      ;;      (org-ml-to-trimmed-string))
-      ;; => (:result "* Footnotes"
-      ;;             "stuff")
-      ;; :end-hidden
+      (:buffer "* dummy"
+               "stuff")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :archivedp t)
+           (org-ml-set-property :commentedp t)
+           (org-ml-set-property :level 2)
+           (org-ml-set-property :pre-blank 1)
+           (org-ml-set-property :priority ?A)
+           (org-ml-set-property :tags '("tmsu"))
+           (org-ml-set-property :title '("smartie"))
+           (org-ml-set-property :todo-keyword "TODO")
+           (org-ml-to-trimmed-string))
+      => (:result "** TODO COMMENT [#A] smartie :tmsu:ARCHIVE:"
+                  ""
+                  "stuff")
+      :begin-hidden
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :footnote-section-p t)
+           (org-ml-to-trimmed-string))
+      => (:result "* Footnotes"
+                  "stuff")
+      :end-hidden
 
-      ;; (:buffer "call_kthulu()")
-      ;; (->> (org-ml-parse-this-object)
-      ;;      (org-ml-set-property :call "cthulhu")
-      ;;      (org-ml-set-property :inside-header '(:cache no))
-      ;;      (org-ml-set-property :arguments '("x=4"))
-      ;;      (org-ml-set-property :end-header '(:exports results))
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "call_cthulhu[:cache no](x=4)[:exports results]"
+      (:buffer "call_kthulu()")
+      (->> (org-ml-parse-this-object)
+           (org-ml-set-property :call "cthulhu")
+           (org-ml-set-property :inside-header '(:cache no))
+           (org-ml-set-property :arguments '("x=4"))
+           (org-ml-set-property :end-header '(:exports results))
+           (org-ml-to-trimmed-string))
+      => "call_cthulhu[:cache no](x=4)[:exports results]"
 
-      ;; (:buffer "src_emacs{(print 'yeah-boi)}")
-      ;; (->> (org-ml-parse-this-object)
-      ;;      (org-ml-set-property :language "python")
-      ;;      (org-ml-set-property :parameters '(:cache no))
-      ;;      (org-ml-set-property :value "print \"yeah boi\"")
-      ;;      (org-ml-to-trimmed-string))
-      ;; => "src_python[:cache no]{print \"yeah boi\"}"
-      ;; :end-hidden
+      (:buffer "src_emacs{(print 'yeah-boi)}")
+      (->> (org-ml-parse-this-object)
+           (org-ml-set-property :language "python")
+           (org-ml-set-property :parameters '(:cache no))
+           (org-ml-set-property :value "print \"yeah boi\"")
+           (org-ml-to-trimmed-string))
+      => "src_python[:cache no]{print \"yeah boi\"}"
+      :end-hidden
 
       (:buffer "- thing")
       (->> (org-ml-parse-this-item)
@@ -1653,6 +1653,60 @@
            (org-ml-set-property :post-blank 1)
            (org-ml-to-string))
       => "*not plain* "
+
+      ;; affiliated keywords
+
+      (:buffer "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :name "foo")
+           (org-ml-to-trimmed-string))
+      => (:result "#+name: foo"
+                  "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :attr_bar '("foo"))
+           (org-ml-to-trimmed-string))
+      => (:result "#+attr_bar: foo"
+                  "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :header '((:k1 "h1") (:k2 "h2")))
+           (org-ml-to-trimmed-string))
+      => (:result "#+header: :k1 h1"
+                  "#+header: :k2 h2"
+                  "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :results '("bar" "foo"))
+           (org-ml-to-trimmed-string))
+      => (:result "#+results[bar]: foo"
+                  "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :caption '("cap"))
+           (org-ml-to-trimmed-string))
+      => (:result "#+caption: cap"
+                  "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :caption '(("foo" "cap")))
+           (org-ml-to-trimmed-string))
+      => (:result "#+caption[foo]: cap"
+                  "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :caption '(("FOO" "CAP") ("foo" "cap")))
+           (org-ml-to-trimmed-string))
+      => (:result "#+caption[FOO]: CAP"
+                  "#+caption[foo]: cap"
+                  "short paragraph")
+      
+      (:buffer "#+caption: cap"
+               "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :caption nil)
+           (org-ml-to-trimmed-string))
+      => "short paragraph"
+      (:buffer "#+name: deleteme"
+               "short paragraph")
+      (->> (org-ml-parse-this-element)
+           (org-ml-set-property :name nil)
+           (org-ml-to-trimmed-string))
+      => "short paragraph"
 
       :end-hidden
 
@@ -1873,7 +1927,6 @@
            (org-ml-get-property :value))
       => "VAL"
 
-      ;; this is stupid, who would ever do this?
       (:buffer "\begin{env}"
                "body"
                "\end{env}")
@@ -1883,7 +1936,6 @@
                   "body"
                   "\end{env}")
 
-      ;; TODO this is also stupid...
       (:buffer "$2+2=4$")
       (->> (org-ml-parse-this-object)
            (org-ml-get-property :value))
@@ -2065,6 +2117,40 @@
       (->> (org-ml-parse-this-object)
            (org-ml-get-property :end))
       => 16
+
+      ;; affiliated keywords
+
+      (:buffer "#+name: name"
+               "#+attr_foo: bar"
+               "#+attr_foo: BAR"
+               "#+plot: poo"
+               "#+caption: koo"
+               "#+caption[COO]: KOO"
+               "#+results[hash]: res"
+               "#+header: :k1 h1"
+               "#+header: :k2 h2"
+               "#+begin_src"
+               "echo test for echo"
+               "#+end_src")
+      (->> (org-ml-parse-this-element)
+           (org-ml-get-property :name))
+      => "name"
+      (->> (org-ml-parse-this-element)
+           (org-ml-get-property :plot))
+      => "poo"
+      (->> (org-ml-parse-this-element)
+           (org-ml-get-property :attr_foo))
+      ;; TODO why are these reversed?
+      => '("BAR" "bar")
+      (->> (org-ml-parse-this-element)
+           (org-ml-get-property :header))
+      => '((:k1 "h1") (:k2 "h2"))
+      (->> (org-ml-parse-this-element)
+           (org-ml-get-property :results))
+      => '("hash" "res")
+      (->> (org-ml-parse-this-element)
+           (org-ml-get-property :caption))
+      => '("koo" ("COO" "KOO"))
       
       :end-hidden